refactor(cli): status, stats, doctor and diagnose move to src/cli/diagnostics.ts
